backend/pyproject.toml
[project]
name = "memrag-backend"
version = "0.1.0"
description = "MemRAG Chatbot Backend - Multimodal AI with RAG & long-term memory"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "google-adk>=1.0.0",
    "google-generativeai>=0.8.0",
    "mem0ai>=0.1.55",
    "qdrant-client>=1.12.0",
    "pypdf>=5.0.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.1",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "langchain-text-splitters>=0.3.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "pyyaml>=6.0",
    "boto3>=1.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "anyio>=4.6.0",
    "pytest-cov>=6.0.0",
    "faker>=30.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
